Adding the initial project
This project is an infographic that is derived from data that is provided, as well as user input data. Object-oriented javascript is great for working with complex data, but it’s not of much use without an interface to interact with. I pulled in information from a form and used it to complete an array of objects that will then be appended back to the DOM.